requirements.txt
face_recognition
Pillow
opencv-python
